merge amir-fix-training-windows into development

+++ criterion/loss.py
@@ -59,14 +59,14 @@ class FocalLoss(nn.Module):
         target = target.view(-1, 1)
 
         logpt = F.log_softmax(input, dim=1)
-        logpt = logpt.gather(1, target)
+        logpt = logpt.gather(1, target.long())
         logpt = logpt.view(-1)
         pt = Variable(logpt.data.exp())
 
         if self.alpha is not None:
             if self.alpha.type() != input.data.type():
                 self.alpha = self.alpha.type_as(input.data)
-            at = self.alpha.gather(0, target.data.view(-1))
+            at = self.alpha.gather(0, target.data.view(-1).long())
             logpt = logpt * Variable(at)
 
         loss = -1 * (1 - pt) ** self.gamma * logpt
